:truck: Move StickerAlbumSimulator to its own file
